change tc_uaf in readings struct to float

+++ controlsoftware/include/serial.h
@@ -19,7 +19,7 @@ struct daqSensors {
 
 struct daqParsed {
 	uint32_t timestamp;
-	int32_t TC_uaf;
+	float TC_uaf;
 	uint8_t HALL_methane, HALL_LOX;
 	float BATT_voltage;
 	int16_t PT_methane, PT_LOX, PT_helium, PT_chamber, PT_heliumReg;
